refactor: Repository Interface 패키징
