+++ src/main/java/gift/config/RepositoryConfig.java
@@ -1,9 +1,7 @@
 package gift.config;
 
-import gift.model.Product;
-import gift.model.ProductRepository;
+import gift.model.repository.ProductRepository;
 import gift.model.Products;
-import gift.model.RepositoryInterface;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
