+++ src/main/java/gift/controller/ProductController.java
@@ -4,7 +4,7 @@ package gift.controller;
 import gift.controller.req.ProductRequest;
 import gift.controller.res.ProductResponse;
 import gift.model.Product;
-import gift.model.ProductRepository;
+import gift.model.repository.ProductRepository;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
