+++ src/main/java/gift/model/repository/ProductRepository.java
@@ -1,4 +1,6 @@
-package gift.model;
+package gift.model.repository;
+
+import gift.model.Product;
 
 public interface ProductRepository extends RepositoryInterface<Product, Long> {
